requirements.txt
python-telegram-bot>=20.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
